Restricción de permisos para poblar la BD, solo el rol Editor puede agregar y modificar la BD para los campos de las preguntas

+++ models/DBmongo/darDatos.js
@@ -1,4 +1,5 @@
 const mongoose = require("mongoose");
+const { getActiveUser } = require("../../authState");
 require("dotenv").config();
 
 const RangoEdad = require("../RangoEdad");
@@ -18,6 +19,13 @@ const connectDB = async () => {
 };
 
 const seedData = async () => {
+  const rolActual = getActiveUser();
+
+  if (rolActual !== "Editor") {
+    console.log("Error: solo el Editor puede poblar la base de datos.");
+    return;
+  }
+
   try {
     console.log("Limpiando datos existentes");
     await Subcategoria.deleteMany({});
